+++ src/components/EditSections/EditUserInfo/components/EditUserProfilePicture/index.js
@@ -1,0 +1,1 @@
+export { default } from './EditUserProfilePicture';
