+++ src/components/EditSections/EditUserInfo/components/ProfilePicture/ProfilePicture.css
@@ -1,0 +1,5 @@
+.profilePlaceholder {
+  max-width: 100px;
+  max-height: 100px;
+  object-fit: scale-down;
+}
