+++ src/components/EditSections/EditUserInfo/components/index.js
@@ -1,4 +1,5 @@
 export { default as ChangeUserTypeModal } from './ChangeUserTypeModal';
 export { default as DeleteProfilePictureModal } from './DeleteProfilePictureModal';
 export { default as ExternalLinkModal } from './ExternalLinkModal';
+export { default as EditUserProfilePicture } from './EditUserProfilePicture';
 export { default as ProfilePicture } from './ProfilePicture';
