+++ src/components/UserDetailSections/UserInfo/UserInfo.js
@@ -1,8 +1,7 @@
 import { get } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage, useIntl } from 'react-intl';
-import { Img } from 'react-image';
+import { FormattedMessage } from 'react-intl';
 import {
   Row,
   Col,
@@ -11,17 +10,14 @@ import {
   Headline,
   NoValue,
   FormattedDate,
-  Loading,
 } from '@folio/stripes/components';
 
 import { ViewMetaData } from '@folio/stripes/smart-components';
 import { useStripes } from '@folio/stripes/core';
-import css from './UserInfo.css';
 import { USER_TYPE_FIELD } from '../../../constants';
-import profilePicThumbnail from '../../../../icons/profilePicThumbnail.png';
-import { isAValidUUID } from '../../util/util';
 
 import { useProfilePicture } from '../../../hooks';
+import ProfilePicture from '../../EditSections/EditUserInfo/components/ProfilePicture';
 
 const UserInfo = (props) => {
   const {
@@ -34,36 +30,13 @@ const UserInfo = (props) => {
   } = props;
   const profilePictureLink = user?.personal?.profilePictureLink;
   const stripes = useStripes();
-  const intl = useIntl();
   const userStatus = (user?.active ?
     <FormattedMessage id="ui-users.active" /> :
     <FormattedMessage id="ui-users.inactive" />);
-  const hasProfilePicture = Boolean(profilePictureLink);
-  /**
-   * Profile Picture Link can be
-   * 1. an id(uuid) of profile picture stored in database or
-   * 2. a link to an image - a url
-   */
-  const isProfilePictureLinkAURL = !isAValidUUID(profilePictureLink);
   const profilePicturesEnabled = Boolean(settings.length) && settings[0].enabled;
   const hasViewProfilePicturePerm = stripes.hasPerm('ui-users.profile-pictures.view');
   const { isFetching, profilePictureData } = useProfilePicture({ profilePictureId: profilePictureLink });
 
-  const renderProfilePic = () => {
-    const profilePictureSrc = isProfilePictureLinkAURL ? profilePictureLink : 'data:;base64,' + profilePictureData;
-    const imgSrc = !hasProfilePicture ? profilePicThumbnail : profilePictureSrc;
-    if (isFetching) {
-      return <Loading />;
-    }
-    return (
-      <Img
-        className={css.profilePlaceholder}
-        alt={intl.formatMessage({ id: 'ui-users.information.profilePicture' })}
-        src={imgSrc}
-        loader={<Loading />}
-      />
-    );
-  };
 
   return (
     <Accordion
@@ -156,7 +129,11 @@ const UserInfo = (props) => {
                 <Col xs={12}>
                   <KeyValue
                     label={<FormattedMessage id="ui-users.information.profilePicture" />}
-                    value={renderProfilePic()}
+                    value={<ProfilePicture
+                      profilePictureLink={profilePictureLink}
+                      isFetching={isFetching}
+                      profilePictureData={profilePictureData}
+                    />}
                   />
                 </Col>
               </Row>
